+++ randomizer-core/src/main/java/de/alpharogroup/random/number/RandomPrimitivesExtensions.java
@@ -24,6 +24,7 @@
  */
 package de.alpharogroup.random.number;
 
+import java.security.SecureRandom;
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
 import java.text.ParseException;
@@ -69,7 +70,24 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static int getRandomIntBetween(int minVolume, int maxVolume)
 	{
-		return minVolume + Math.abs(DefaultSecureRandom.get().nextInt()) % maxVolume;
+		return getRandomIntBetween(minVolume, maxVolume, DefaultSecureRandom.get());
+	}
+
+	/**
+	 * Returns a random int between the range from minVolume and maxVolume with the
+	 * <code>Math.abs</code> method.
+	 *
+	 * @param minVolume
+	 *            the min volume
+	 * @param maxVolume
+	 *            the max volume
+	 * @param secureRandom
+	 *            the secure random for number generation
+	 * @return A random int between the range from minVolume and maxVolume
+	 */
+	public static int getRandomIntBetween(int minVolume, int maxVolume, SecureRandom secureRandom)
+	{
+		return minVolume + Math.abs(Objects.requireNonNull(secureRandom).nextInt()) % maxVolume;
 	}
 
 	/**
@@ -83,6 +101,17 @@ public final class RandomPrimitivesExtensions
 	}
 
 	/**
+	 * Returns a random boolean.
+	 *
+	 * @return The random boolean.
+	 */
+	public static boolean randomBoolean(SecureRandom secureRandom)
+	{
+		return randomInt(2, RandomAlgorithm.SECURE_RANDOM,
+			Objects.requireNonNull(secureRandom)) == 0;
+	}
+
+	/**
 	 * The Method randomByte() selects a random byte.
 	 *
 	 * @return The random byte.
@@ -90,6 +119,17 @@ public final class RandomPrimitivesExtensions
 	public static byte randomByte()
 	{
 		return (byte)randomInt(255);
+	}
+
+	/**
+	 * The Method randomByte() selects a random byte.
+	 *
+	 * @return The random byte.
+	 */
+	public static byte randomByte(SecureRandom secureRandom)
+	{
+		return (byte)randomInt(255, RandomAlgorithm.SECURE_RANDOM,
+			Objects.requireNonNull(secureRandom));
 	}
 
 	/**
@@ -101,6 +141,19 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static byte[] randomByteArray(final int length)
 	{
+		return randomByteArray(length, DefaultSecureRandom.get());
+	}
+
+	/**
+	 * The Method randomByteArray(int) generates a random byte array.
+	 *
+	 * @param length
+	 *            the length.
+	 * @return the byte[]
+	 */
+	public static byte[] randomByteArray(final int length, SecureRandom secureRandom)
+	{
+		Objects.requireNonNull(secureRandom);
 		final byte[] randomByteArray = new byte[length];
 		final byte[] randomByteBox = new byte[1];
 		for (int i = 0; i < length; i++)
@@ -111,7 +164,7 @@ public final class RandomPrimitivesExtensions
 			}
 			else
 			{
-				DefaultSecureRandom.get().nextBytes(randomByteBox);
+				secureRandom.nextBytes(randomByteBox);
 				randomByteArray[i] = randomByteBox[0];
 			}
 		}
@@ -125,15 +178,29 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static char randomChar()
 	{
-		if (DefaultSecureRandom.get().nextBoolean())
+		return randomChar(DefaultSecureRandom.get());
+	}
+
+	/**
+	 * Returns a random char.
+	 * 
+	 * @param secureRandom
+	 *            the secure random for char generation
+	 *
+	 * @return The generated random char.
+	 */
+	public static char randomChar(SecureRandom secureRandom)
+	{
+		Objects.requireNonNull(secureRandom);
+		if (secureRandom.nextBoolean())
 		{
 			// random character
-			return (char)(DefaultSecureRandom.get().nextInt(26) + 65);
+			return (char)(secureRandom.nextInt(26) + 65);
 		}
 		else
 		{
 			// random digit
-			return (char)DefaultSecureRandom.get().nextInt(10);
+			return (char)secureRandom.nextInt(10);
 		}
 	}
 
@@ -185,11 +252,31 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static double randomDouble(final double range, final RandomAlgorithm algorithm)
 	{
+		return randomDouble(range, Objects.requireNonNull(algorithm), DefaultSecureRandom.get());
+	}
+
+	/**
+	 * Gets an random double to the given range with the given random algorithm <br>
+	 * <br>
+	 * For example: if you put range to 10 the random float is between 0.0-9.9
+	 *
+	 * @param range
+	 *            the range
+	 * @param algorithm
+	 *            the random algorithm
+	 * @param secureRandom
+	 *            the secure random for double generation
+	 * @return an random double not greater then the range
+	 */
+	public static double randomDouble(final double range, final RandomAlgorithm algorithm,
+		SecureRandom secureRandom)
+	{
 		Objects.requireNonNull(algorithm);
+		Objects.requireNonNull(secureRandom);
 		switch (algorithm)
 		{
 			case MATH_ABS :
-				return Math.abs(DefaultSecureRandom.get().nextDouble()) % range;
+				return Math.abs(secureRandom.nextDouble()) % range;
 			case MATH_RANDOM :
 				return Math.random() * range;
 			case RANDOM :
@@ -197,7 +284,7 @@ public final class RandomPrimitivesExtensions
 				return MathExtensions.isPositive(random) ? random : random * -1;
 			case SECURE_RANDOM :
 			default :
-				return DefaultSecureRandom.get().nextDouble() * range;
+				return secureRandom.nextDouble() * range;
 		}
 	}
 
@@ -245,7 +332,7 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static float randomFloat()
 	{
-		return randomFloat(DefaultSecureRandom.get().nextFloat());
+		return randomFloat(DefaultSecureRandom.get());
 	}
 
 	/**
@@ -274,11 +361,31 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static float randomFloat(final float range, final RandomAlgorithm algorithm)
 	{
+		return randomFloat(range, Objects.requireNonNull(algorithm), DefaultSecureRandom.get());
+	}
+
+	/**
+	 * Gets an random float to the given range with the given random algorithm <br>
+	 * <br>
+	 * For example: if you put range to 10 the random float is between 0.0-9.9
+	 *
+	 * @param range
+	 *            the range
+	 * @param algorithm
+	 *            the random algorithm
+	 * @param secureRandom
+	 *            the secure random for float generation
+	 * @return an random float not greater then the range
+	 */
+	public static float randomFloat(final float range, final RandomAlgorithm algorithm,
+		SecureRandom secureRandom)
+	{
 		Objects.requireNonNull(algorithm);
+		Objects.requireNonNull(secureRandom);
 		switch (algorithm)
 		{
 			case MATH_ABS :
-				return (float)(Math.abs(DefaultSecureRandom.get().nextDouble()) % range);
+				return (float)(Math.abs(secureRandom.nextDouble()) % range);
 			case MATH_RANDOM :
 				return (float)(Math.random() * range);
 			case RANDOM :
@@ -286,7 +393,7 @@ public final class RandomPrimitivesExtensions
 				return MathExtensions.isPositive(random) ? random : random * -1;
 			case SECURE_RANDOM :
 			default :
-				return (float)(DefaultSecureRandom.get().nextDouble() * range);
+				return (float)(secureRandom.nextDouble() * range);
 		}
 	}
 
@@ -302,6 +409,19 @@ public final class RandomPrimitivesExtensions
 	public static float randomFloat(final int afterComma, final int beforeComma)
 	{
 		return Float.parseFloat(getRandomFloatString(afterComma, beforeComma));
+	}
+
+	/**
+	 * Generates a random float between the range 0.0-9.9.
+	 * 
+	 * @param secureRandom
+	 *            the secure random for float generation
+	 *
+	 * @return the generated random float between the range 0.0-9.9.
+	 */
+	public static float randomFloat(SecureRandom secureRandom)
+	{
+		return randomFloat(Objects.requireNonNull(secureRandom).nextFloat());
 	}
 
 	/**
@@ -348,7 +468,7 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static int randomInt()
 	{
-		return randomInt(DefaultSecureRandom.get().nextInt());
+		return randomInt(DefaultSecureRandom.get());
 	}
 
 	/**
@@ -376,11 +496,30 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static int randomInt(final int range, final RandomAlgorithm algorithm)
 	{
+		return randomInt(range, Objects.requireNonNull(algorithm), DefaultSecureRandom.get());
+	}
+
+	/**
+	 * The Method randomInt(int) gets an int to the spezified range. For example: if you put range
+	 * to 10 the random int is between 0-9.
+	 *
+	 * @param range
+	 *            The range
+	 * @param algorithm
+	 *            the random algorithm
+	 * @param secureRandom
+	 *            the secure random for int generation
+	 * @return an int not greater then the range
+	 */
+	public static int randomInt(final int range, final RandomAlgorithm algorithm,
+		SecureRandom secureRandom)
+	{
 		Objects.requireNonNull(algorithm);
+		Objects.requireNonNull(secureRandom);
 		switch (algorithm)
 		{
 			case MATH_ABS :
-				return Math.abs(DefaultSecureRandom.get().nextInt()) % range;
+				return Math.abs(secureRandom.nextInt()) % range;
 			case MATH_RANDOM :
 				return (int)(Math.random() * range);
 			case RANDOM :
@@ -388,8 +527,21 @@ public final class RandomPrimitivesExtensions
 				return MathExtensions.isPositive(random) ? random : random * -1;
 			case SECURE_RANDOM :
 			default :
-				return (int)(DefaultSecureRandom.get().nextDouble() * range);
+				return (int)(secureRandom.nextDouble() * range);
 		}
+	}
+
+	/**
+	 * The Method randomInt() gets an int between the range 0-9.
+	 * 
+	 * @param secureRandom
+	 *            the secure random for int generation
+	 *
+	 * @return an int between the range 0-9.
+	 */
+	public static int randomInt(SecureRandom secureRandom)
+	{
+		return randomInt(Objects.requireNonNull(secureRandom).nextInt());
 	}
 
 	/**
@@ -445,7 +597,7 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static long randomLong()
 	{
-		return randomLong(new Random(System.currentTimeMillis()).nextInt());
+		return randomLong(DefaultSecureRandom.get());
 	}
 
 	/**
@@ -474,11 +626,29 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static long randomLong(final long range, final RandomAlgorithm algorithm)
 	{
+		return randomLong(range, Objects.requireNonNull(algorithm), DefaultSecureRandom.get());
+	}
+
+	/**
+	 * Gets an random long to the given range with the given random algorithm <br>
+	 * <br>
+	 * For example: if you put range to 10 the random int is between 0-9
+	 *
+	 * @param range
+	 *            the range
+	 * @param algorithm
+	 *            the random algorithm
+	 * @return an random long not greater then the range
+	 */
+	public static long randomLong(final long range, final RandomAlgorithm algorithm,
+		SecureRandom secureRandom)
+	{
 		Objects.requireNonNull(algorithm);
+		Objects.requireNonNull(secureRandom);
 		switch (algorithm)
 		{
 			case MATH_ABS :
-				return (long)(Math.abs(DefaultSecureRandom.get().nextDouble()) % range);
+				return (long)(Math.abs(secureRandom.nextDouble()) % range);
 			case MATH_RANDOM :
 				return (long)(Math.random() * range);
 			case RANDOM :
@@ -486,8 +656,21 @@ public final class RandomPrimitivesExtensions
 				return MathExtensions.isPositive(random) ? random : random * -1;
 			case SECURE_RANDOM :
 			default :
-				return (long)(DefaultSecureRandom.get().nextDouble() * range);
+				return (long)(secureRandom.nextDouble() * range);
 		}
+	}
+
+	/**
+	 * Gets a random long
+	 * 
+	 * @param secureRandom
+	 *            the secure random for long generation
+	 *
+	 * @return a random long
+	 */
+	public static long randomLong(SecureRandom secureRandom)
+	{
+		return randomLong(Objects.requireNonNull(secureRandom).nextLong());
 	}
 
 	/**
@@ -511,13 +694,24 @@ public final class RandomPrimitivesExtensions
 	 */
 	public static short randomShort()
 	{
-		if (DefaultSecureRandom.get().nextBoolean())
+		return randomShort(DefaultSecureRandom.get());
+	}
+
+	/**
+	 * Returns a random short
+	 *
+	 * @return The generated random short
+	 */
+	public static short randomShort(SecureRandom secureRandom)
+	{
+		Objects.requireNonNull(secureRandom);
+		if (secureRandom.nextBoolean())
 		{
-			return (short)(DefaultSecureRandom.get().nextInt(65536) - 32768);
+			return (short)(secureRandom.nextInt(65536) - 32768);
 		}
 		else
 		{
-			return (short)DefaultSecureRandom.get().nextInt(Short.MAX_VALUE + 1);
+			return (short)secureRandom.nextInt(Short.MAX_VALUE + 1);
 		}
 	}